+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.9.7]
 ### Added
 - Detect anomalies resulting from dependency violations
+- Add intersecting label equivalence relation
 
 ## [0.9.6]
 ### Fixed